Use new component structure
