+++ src/Price/Price.spec.js
